+++ src/main/java/svenhjol/charmony/runestones/RunestonesMod.java
@@ -1,9 +1,8 @@
 package svenhjol.charmony.runestones;
 
-import net.minecraft.resources.ResourceLocation;
 import svenhjol.charmony.api.core.ModDefinition;
-import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.api.core.Side;
+import svenhjol.charmony.core.base.Mod;
 
 @ModDefinition(
     id = RunestonesMod.ID,
@@ -22,9 +21,5 @@ public final class RunestonesMod extends Mod {
             instance = new RunestonesMod();
         }
         return instance;
-    }
-
-    public static ResourceLocation id(String path) {
-        return ResourceLocation.fromNamespaceAndPath(ID, path);
     }
 }